misc stuff
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@5468 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Common/Src/SDCardUtil.cpp
@@ -186,7 +186,7 @@ static int write_empty(FILE* file, u64 count)
 	return 0;
 }
 
-bool SDCardCreate(u32 disk_size /*in MB*/, char* filename)
+bool SDCardCreate(u64 disk_size /*in MB*/, char* filename)
 {
 	int sectors_per_fat;
 	int sectors_per_disk;
